feat: 게시글 정보에 작성자 id 저장하도록 변경

+++ src/deskteriors/deskteriors.service.ts
@@ -3,6 +3,7 @@ import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Deskterior } from './deskterior.entity';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class DeskteriorsService {
@@ -28,12 +29,17 @@ export class DeskteriorsService {
   }
 
   /** 데스크테리어 글 작성 */
-  async createBoard(deskteriorDto: DeskteriorDto): Promise<Deskterior> {
+  async createBoard(
+    deskteriorDto: DeskteriorDto,
+    user: User,
+  ): Promise<Deskterior> {
     const { title, description } = deskteriorDto;
 
     const new_board = this.deskteriorRepository.create({
       title,
       description,
+      user,
+      authorId: user.id,
     });
 
     await this.deskteriorRepository.save(new_board);
@@ -41,20 +47,34 @@ export class DeskteriorsService {
   }
 
   /** 데스크테리어 글 삭제 */
-  async deleteBoardById(id: number): Promise<void> {
-    await this.deskteriorRepository.delete({ id });
+  async deleteBoardById(
+    id: number,
+    user: User,
+  ): Promise<void | { message: string }> {
+    const found = await this.findOneById(id);
+
+    if (found.authorId === user.id) {
+      await this.deskteriorRepository.delete({ id });
+    } else {
+      return { message: '권한이 없습니다.' };
+    }
   }
 
   /** 데스크테리어 글 수정 */
   async updateBoardById(
     id: number,
     newDescription: string,
-  ): Promise<Deskterior> {
+    user: User,
+  ): Promise<Deskterior | { message: string }> {
     const found = await this.findOneById(id);
 
-    found.description = newDescription;
-    this.deskteriorRepository.save(found);
+    if (found.authorId === user.id) {
+      found.description = newDescription;
+      this.deskteriorRepository.save(found);
 
-    return found;
+      return found;
+    } else {
+      return { message: '권한이 없습니다.' };
+    }
   }
 }
